test numeric for sel_grid

+++ src/functions.R
@@ -210,7 +210,7 @@ make_sel_grid <- function(df, outcome_type = NULL, protocol = NULL) {
         ttfu = unique(df$bin_ttfu),
         rob = unique(df$bin_rob)
     )
-
+    # TODO: REMOVE
     ##### Drop doctreat when outcome = "func"
     if (outcome_type == "func") {
         grid_vals[["doctreat"]] <- 2
@@ -267,7 +267,7 @@ make_sel_grid <- function(df, outcome_type = NULL, protocol = NULL) {
     ### CREATE GRID
     grid <- expand_grid.(!!!grid_vals) %>%
         mutate.(
-            across.(where(is.numeric), as.integer),
+            across.(where(is.integer), as.numeric),
             across.(where(is.character), as_factor)
         )
